Normalize neutronics integral scalars for strict mypy

+++ src/scpn_fusion/nuclear/blanket_neutronics.py
@@ -523,7 +523,9 @@ class MultiGroupBlanket:
         total_prod = prod_g1 + prod_g2 + prod_g3
 
         if hasattr(np, "trapezoid"):
-            total_tritium = np.trapezoid(total_prod * 2.0 * np.pi * self.r, self.r)
+            total_tritium = float(
+                np.trapezoid(total_prod * 2.0 * np.pi * self.r, self.r)
+            )
         else:  # pragma: no cover - legacy NumPy compatibility path
             edge = np.diff(self.r)
             total_integrand = total_prod * 2.0 * np.pi * self.r
@@ -539,9 +541,9 @@ class MultiGroupBlanket:
         # Per-group TBR breakdown (with same correction factors as total)
         corr = port_coverage_factor * streaming_factor
         if hasattr(np, "trapezoid"):
-            tbr_g1_raw = np.trapezoid(prod_g1 * 2.0 * np.pi * self.r, self.r)
-            tbr_g2_raw = np.trapezoid(prod_g2 * 2.0 * np.pi * self.r, self.r)
-            tbr_g3_raw = np.trapezoid(prod_g3 * 2.0 * np.pi * self.r, self.r)
+            tbr_g1_raw = float(np.trapezoid(prod_g1 * 2.0 * np.pi * self.r, self.r))
+            tbr_g2_raw = float(np.trapezoid(prod_g2 * 2.0 * np.pi * self.r, self.r))
+            tbr_g3_raw = float(np.trapezoid(prod_g3 * 2.0 * np.pi * self.r, self.r))
         else:  # pragma: no cover - legacy NumPy compatibility path
             edge = np.diff(self.r)
             g1 = prod_g1 * 2.0 * np.pi * self.r
